move pair selection by query params to trade page instead of pair selector component
